Moved Edit/Delete buttons to seperate component
