+++ components/FoodItem.js
@@ -1,20 +1,6 @@
 import React from 'react';
-import { View, StyleSheet } from 'react-native';
-import { List, FAB } from 'react-native-paper';
-
-const EditDelete = ({props}) => {
-    const {navigation, id, type, deleteFood, calories, name} = props;
-    return (
-        <View style={styles.flex}>
-            <FAB 
-                style={styles.fab}
-                small
-                icon="pencil"
-                onPress={() => navigation.navigate('Form', {formType: 'edit', mealType: type, id, calories, title:name})}/>
-            <FAB style={styles.fab} small icon='close' onPress={() => deleteFood(id)} />
-        </View>
-    )
-}
+import {List} from 'react-native-paper';
+import EditDelete from './EditDelete';
 
 export const FoodItem = (props) => {
     return (
@@ -24,15 +10,3 @@ export const FoodItem = (props) => {
         />
     )
 };
-
-const styles = StyleSheet.create({
-    flex: {
-        flexDirection: 'row'
-
-    },
-    fab: {
-        margin: 5,
-        right: 0,
-        bottom: 0
-    }
-})
